Split out isFeaturedImageInContent
