+++ client/lib/post-normalizer/utils/index.js
@@ -1,15 +1,3 @@
-/**
- * External Dependencies
- */
-import { findIndex } from 'lodash';
-import url from 'url';
-
-/**
- * Internal Dependencies
- */
-import { isPhotonHost } from 'lib/post-normalizer/utils/is-photon-host';
-import { thumbIsLikelyImage } from 'lib/post-normalizer/utils/thumb-is-likely-image';
-
 export { isPhotonHost } from 'lib/post-normalizer/utils/is-photon-host';
 export { imageSizeFromAttachments } from 'lib/post-normalizer/utils/image-size-from-attachments';
 export { maxWidthPhotonishURL } from 'lib/post-normalizer/utils/max-width-photonish-url';
@@ -19,39 +7,7 @@ export { isUrlLikelyAnImage } from 'lib/post-normalizer/utils/is-url-likely-an-i
 export { thumbIsLikelyImage } from 'lib/post-normalizer/utils/thumb-is-likely-image';
 export { iframeIsWhitelisted } from 'lib/post-normalizer/utils/iframe-is-whitelisted';
 export { isCandidateForCanonicalImage } from 'lib/post-normalizer/utils/is-candidate-for-canonical-image';
-
-function getPathname( uri ) {
-	const parsedUrl = url.parse( uri );
-	const path = parsedUrl.pathname;
-	if ( isPhotonHost( parsedUrl.hostname ) ) {
-		return path.substring( path.indexOf( '/', 1 ) );
-	}
-	return path;
-}
-
-/** returns whether or not a posts featuredImages is contained within the contents
- *
- * @param {object} post - the post to check
- * @returns {boolean|number} false if featuredImage is not within content content_images.
- *   otherwise returns the index of the dupe in post.images.
- */
-export function isFeaturedImageInContent( post ) {
-	if ( thumbIsLikelyImage( post.post_thumbnail ) ) {
-		const featuredImagePath = getPathname( post.post_thumbnail.URL );
-
-		const indexOfContentImage = findIndex(
-			post.images,
-			img => getPathname( img.src ) === featuredImagePath,
-			1
-		); // skip first element in post.images because it is always the featuredImage
-
-		if ( indexOfContentImage > 0 ) {
-			return indexOfContentImage;
-		}
-	}
-
-	return false;
-}
+export { isFeaturedImageInContent } from 'lib/post-normalizer/utils/is-featured-image-in-content';
 
 export function deduceImageWidthAndHeight( image ) {
 	if ( image.height && image.width ) {
